Refactor Placed requests
