+++ frontend/src/components/requester/my_requests/placed_request/request_images.module.css
@@ -1,0 +1,22 @@
+.imagesContainer {
+    display: grid;
+    margin-top: 1em;
+    width: min(90vw, 550px);
+    gap: 0.5em;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+}
+.singleImage {
+    width: fit-content;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 0.3em;
+}
+
+.singleImage img {
+    height: auto;
+    box-shadow: 0px 0px 12px rgba(0, 0, 0, 0.137);
+
+    border-radius: 8px;
+    width: clamp(160px, 50%, 200px);
+}
